// src/app/(protected)/master-data/campaigns/types.d.ts
export type CampaignsItem = {
	id: number
	countryImageSrc: string
	fillingsDate: string
	closing: string
	activist: string
	activistImageSrc: string
	target: string
	targetImageSrc: string
	objective: string
	marketCap: number
	positionSize: number
	stake: number
	return: number
	abnormalReturn: number
	proxyContestDate: string
	settlementDate: string
	outcome: string
}